Separate IterationPolice from SplayIterator, cause it can be used in both trees
