+++ src/SplayIterator.hpp
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "IterationPolicy.h"
+
 #include <memory>
 
 template<typename T>
@@ -7,51 +9,6 @@ class Tree;
 
 template <typename T>
 class SplayTree;
-
-
-
-template<typename T>
-class IterationPolicy{
-    /**
-     *
-     * Base class for iteration policies
-     * @brief: Defines behaviour of iterator using IterationPolicy::next(node)
-     *
-     * */
-public:
-    virtual std::shared_ptr<typename Tree<T>::Node> next(std::shared_ptr<typename Tree<T>::Node> node) const noexcept = 0;
-};
-
-
-
-template<typename T>
-class ForwardIteration : public IterationPolicy<T>{
-    /**
-     *
-     * Extends IterationPolicy
-     * @brief: Allows us to iterate on tree in forward order
-     *
-     * */
-public:
-    std::shared_ptr<typename Tree<T>::Node> next(std::shared_ptr<typename Tree<T>::Node> node) const noexcept override;
-};
-
-
-
-template<typename T>
-class ReverseIteration : public IterationPolicy<T>{
-    /**
-     *
-     * Extends IterationPolicy
-     * @brief: Allows us to iterate on tree in reverse order
-     *
-     * */
-public:
-    std::shared_ptr<typename Tree<T>::Node> next(std::shared_ptr<typename Tree<T>::Node> node) const noexcept override;
-};
-
-
-
 
 template<typename T>
 class SplayIterator {
